% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_lifecycle_badges.R
\name{copy_lifecycle_figures}
\alias{copy_lifecycle_figures}
\title{Copy into ./man/files the figures required to support lifecycle badges.}
\usage{
copy_lifecycle_figures(new_path = here::here("man", "figures"))
}
\arguments{
\item{new_path}{Destination path of the copied figures.}
}
\value{
Invisible \code{new_path}.
}
\description{
\Sexpr[results=rd, stage=render]{fgeo.template:::lifecycle("maturing")}
}
\details{
To support lifecycle badges you need to first run
\code{use_template("compat-lifecycle.R", package = "fgeo.template")}. See also
\url{https://github.com/r-lib/rlang/blob/master/R/compat-lifecycle.R}.
}
\examples{
\dontrun{
# This creates a file in the root of your package. Move it to R/
use_template("compat-lifecycle.R", package = "fgeo.template")
# This adds the images for the badges into man/figures
copy_lifecycle_figures()
}
}
