% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.template-package.R
\docType{package}
\name{fgeo.template-package}
\alias{fgeo.template}
\alias{fgeo.template-package}
\title{fgeo.template: Templates Used to Build ForestGEO Packages}
\description{
Templates used to build ForestGEO packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.template}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.template/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
