% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_lifecycle_badges.R
\name{lifecycle}
\alias{lifecycle}
\title{Embed a lifecycle badge in documentation}
\arguments{
\item{stage}{A lifecycle stage as a string, one of:
\code{"experimental"}, \code{"maturing"}, \code{"stable"}, \code{"questioning"},
\code{"archived"}, \code{"soft-deprecated"}, \code{"deprecated"}, \code{"defunct"}.}
}
\description{
(From \url{https://github.com/r-lib/rlang/blob/master/R/compat-lifecycle.R})

Use \code{lifecycle()} within a \code{Sexpr} macro to embed a
\href{https://www.tidyverse.org/lifecycle/}{lifecycle} badge in your
documentation. The badge should appear first in the description:\preformatted{\Sexpr[results=rd, stage=render]{mypkg:::lifecycle("questioning")}
}

The badge appears as an image in the HTML version of the
documentation. To make them available in your package, visit
\url{https://github.com/r-lib/rlang/tree/master/man/figures} and copy
all the files starting with \code{lifecycle-} in your \code{man/figures/}
folder.
}
